% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\alias{with_headers}
\title{Create an API call}
\usage{
api(url, headers = NULL, response = "text")

with_headers(...)
}
\arguments{
\item{url}{the url to request}

\item{headers}{optional headers for the call}

\item{response}{the type of response expected}

\item{...}{headers converted into a list}
}
\value{
a \code{trapi} object
}
\description{
Create an API call
}
\section{Functions}{
\itemize{
\item \code{with_headers}: Api Headers
}}

\author{
Tom Elliott
}
