% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{construct}
\alias{construct}
\title{Construct transit network}
\usage{
construct(nw)
}
\arguments{
\item{nw}{a \code{trgtfs} object}
}
\value{
A \code{trgtfs} object, with additional network tables.
}
\description{
Constructs a transit network from existing GTFS shapes.
Detects intersections and roads between them.
}
\author{
Tom Elliott
}
