% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{create_gtfs}
\alias{create_gtfs}
\title{Create GTFS database}
\usage{
create_gtfs(source, db = tempfile(), quiet = FALSE,
  output = "predictions.pb")
}
\arguments{
\item{source}{where the GTFS data comes from}

\item{db}{where the GTFS data is going to be stored}

\item{quiet}{logical, if \code{TRUE} progress output will be suppressed}

\item{output}{filename to output ETA predictions}
}
\value{
a \code{trgtfs} object
}
\description{
Create a new GTFS database, optionally from a path.
}
\details{
The GTFS data, if specified, should be specifided as a series of text files.
See \code{update} for more info.
}
\author{
Tom Elliott
}
