% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run realtime transit model}
\usage{
model(nw, n.particles = 500, cores = 1L)
}
\arguments{
\item{nw}{a loaded \code{trgtfs} object}

\item{n.particles}{the number of particles to use to model each vehicle}

\item{cores}{the number of cores to use in parallel}
}
\value{
NULL (in fact, it never returns anything ...)
}
\description{
Run realtime transit model
}
\author{
Tom Elliott
}
