% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime_feed.R
\name{realtime_feed}
\alias{realtime_feed}
\title{Connect to GTFS feed}
\usage{
realtime_feed(nw, url, headers = NULL, response = c("json", "protobuf"))
}
\arguments{
\item{nw}{a \code{trgtfs} object}

\item{url}{the API url}

\item{headers}{a header object from \code{with_headers} defining any GET request headers (e.g., api keys)}

\item{response}{the type of response expected (json or protobuf)}
}
\value{
a \code{trgtfs} object with feed info
}
\description{
Connect to a realtime GTFS feed
}
\author{
Tom Elliott
}
