% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GtfsRoute.R
\name{route}
\alias{route}
\alias{show,Rcpp_Route-method}
\title{Gtfs 'Route' Constructor}
\usage{
route(
  id,
  short_name = "",
  long_name = "",
  type = 0L,
  agency_id = "",
  version = 0
)

\S4method{show}{Rcpp_Route}(object)
}
\arguments{
\item{id}{Route ID}

\item{short_name}{Route short name}

\item{long_name}{Route long name}

\item{type}{Route type}

\item{agency_id}{Agency ID}

\item{version}{GTFS Version}
}
\value{
Rcpp_Route object
}
\description{
Gtfs 'Route' Constructor
}
\section{Functions}{
\itemize{
\item \code{show(Rcpp_Route)}: Show method

}}
