% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters for the model}
\usage{
set_parameters(nw, ...)
}
\arguments{
\item{nw}{a loaded \code{trgtfs} object}

\item{...}{parameters}
}
\description{
Set parameters for the model
}
\author{
Tom Elliott
}
