% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitr.R
\docType{package}
\name{transitr}
\alias{transitr}
\title{transitr}
\description{
Package for working with transit GTFS data
}
\author{
Tom Elliott \href{mailto:tom.elliott@auckland.ac.nz}{tom.elliott@auckland.ac.nz}
}
