% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GtfsTrip.R
\name{trip}
\alias{trip}
\alias{show,Rcpp_Trip-method}
\title{GTFS Trip constructor}
\usage{
trip(id, route, direction_id = 0, headsign = "", version = 0)

\S4method{show}{Rcpp_Trip}(object)
}
\arguments{
\item{id}{Trip ID}

\item{route}{Route object, constructed using \code{route()}}

\item{direction_id}{Direction ID (0 or 1)}

\item{headsign}{Trip headsign}

\item{version}{GTFS Version}
}
\value{
Rcpp_Trip object
}
\description{
GTFS Trip constructor
}
\section{Functions}{
\itemize{
\item \code{show(Rcpp_Trip)}: Show method

}}
