% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_realtime.R
\name{view_realtime}
\alias{view_realtime}
\title{View Realtime Vehicle States}
\usage{
view_realtime(db)
}
\arguments{
\item{db}{SQLite database containing vehicle states, ETAs, etc}
}
\description{
View the realtime state of vehicles
}
\author{
Tom Elliott
}
