% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{create_gtfs}
\alias{create_gtfs}
\title{Create GTFS database}
\usage{
create_gtfs(from, to = tempfile())
}
\arguments{
\item{from}{where the GTFS data comes from}

\item{to}{where the GTFS data is going to be stored}
}
\value{
a \code{trgtfs} object
}
\description{
Create a new GTFS database, optionally from a path.
}
\details{
The GTFS data, if specified, should be specifided as a series of text files.
See \code{update} for more info.
}
\author{
Tom Elliott
}
