% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{load_gtfs}
\alias{load_gtfs}
\title{Load GTFS database}
\usage{
load_gtfs(con)
}
\arguments{
\item{con}{a database connection from \code{dbConnect}}
}
\value{
a \code{trgtfs} object
}
\description{
Load an existing GTFS database into R for use with transitr.
}
\author{
Tom Elliott
}
