% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Run realtime transit model}
\usage{
model(nw)
}
\arguments{
\item{nw}{a loaded \code{trgtfs} object}
}
\value{
NULL (in fact, it never returns anything ...)
}
\description{
Run realtime transit model
}
\author{
Tom Elliott
}
