% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{update.trgtfs}
\alias{update.trgtfs}
\title{Update GTFS data}
\usage{
\method{update}{trgtfs}(object, from)
}
\arguments{
\item{object}{the \code{trgtfs} database object to update}

\item{from}{where the new data should come from}
}
\value{
a \code{trgtfs} object (with updated data)
}
\description{
Update a GTFS database object with new data.
}
\details{
The data is expected to consist of a series of GTFS text/csv files,
and can be specified as either
\begin{itemize}
  \item a URL to zipped GTFS data
  \item a path to a zipped GTFS file
  \item a path to a GTFS directory of text files
\end{itemize}
}
\author{
Tom Elliott
}
